pyproject.toml
[tool.poetry]
name = "ivymart_backend"
version = "0.1.0"
description = ""
authors = ["Ankush <ankush357159@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.13.1"
Pillow = "^9.0.0"
djangorestframework-simplejwt = "^5.0.0"
django-cors-headers = "^3.11.0"
django-cloudinary-storage = "^0.3.0"
stripe = "^2.65.0"
django-stripe-payments = "^2.0.0"
razorpay = "^1.3.0"
django-environ = "^0.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
